refactor(cli): decompose rex/cli.py into rex/commands domain modules (US-REM-027)
rex/cli.py now keeps parser registration, main(), and backward-compatible re-exports while handlers and parser blocks moved verbatim into 13 domain modules. Help output, command behavior, and rex.cli patch points are unchanged.
