+++ rex/commands/__init__.py
@@ -1,0 +1,1 @@
+"""Rex CLI command modules, decomposed by domain (US-REM-027)."""
